Prevented script failures in case they ran already and disks were persistent

+++ specs/master/cluster-init/scripts/10.install_singularity.sh
@@ -1,9 +1,14 @@
 #!/bin/bash
 
+set -ex
+
 # changes here need to be reflected in the script that adds to PATH
 VERSION=3.4.0
 PREFIX=/shared
 #PREFIX=/mnt/resource/apps/singularity-${VERSION}; local only
+
+test -e $PREFIX/bin/singularity && exit 0
+
 yum -y install golang openssl-devel libuuid-devel
 wget https://github.com/sylabs/singularity/releases/download/v${VERSION}/singularity-${VERSION}.tar.gz
 tar -xzf singularity-${VERSION}.tar.gz
