+++ specs/master/cluster-init/scripts/30.install_azcopy.sh
@@ -4,6 +4,8 @@ set -ex
 
 PREFIX=/shared
 
+test -e $PREFIX/bin/azcopy && exit 0 
+
 cd $(mktemp -d)
 wget -O azcopy.tar.gz https://aka.ms/downloadazcopy-v10-linux
 tar -xf azcopy.tar.gz
